requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.3
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-tinymce4-lite==1.7.5
djangorestframework==3.10.3
idna==2.8
jsmin==2.2.2
python-slugify==3.0.3
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.5
